+++ lib/src/client/cargo.dart
@@ -6,7 +6,7 @@ const String CARGO_MODE_LOCAL = "localstorage";
 abstract class Cargo extends CargoBase with CargoDispatch {
   Cargo._();
   /// Create a new cargo storage
-  factory Cargo({CargoModeHolder MODE: CargoMode.MEMORY, String path: null}) {
+  factory Cargo({CargoModeHolder MODE: CargoMode.MEMORY, Map path}) {
       print("Initiating a cargo storage with ${MODE} backend");
       
       switch(MODE) {